% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filelock-package.R
\docType{package}
\name{filelock-package}
\alias{filelock}
\alias{filelock-package}
\title{filelock: Portable File Locking}
\description{
Place an exclusive or shared lock on a file. It uses 'LockFile' on Windows and 'fcntl' locks on Unix-like systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-lib.github.io/filelock/}
  \item \url{https://github.com/r-lib/filelock}
  \item Report bugs at \url{https://github.com/r-lib/filelock/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
