% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{filelock}
\alias{filelock}
\alias{filelock-package}
\title{Portable File Locking}
\description{
Place an exclusive or shared lock on a file. It uses \code{LockFile} on
Windows and \code{fcntl} locks on Unix-like systems.
}
