% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lock}
\alias{lock}
\alias{unlock}
\title{File locking and unlocking}
\usage{
lock(path, exclusive = TRUE, timeout = Inf)

unlock(lock)
}
\arguments{
\item{path}{Path to the file to lock.}

\item{exclusive}{Whether to acquire an exclusive lock.}

\item{timeout}{Timeout to acquire the lock in milliseconds. If \code{Inf},
then the process will wait indefinitely to acquire the lock. If zero,
then it returns immediately.}

\item{lock}{The lock object to unlock.}
}
\description{
TODO: details.
}
\examples{
# TODO
}
